requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
Authlib==1.6.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==45.0.5
cycler==0.12.1
cyclopts==3.22.2
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
dnspython==2.7.0
docstring_parser==0.16
docutils==0.21.2
duckdb==1.3.1
email_validator==2.2.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.8
fastmcp==2.10.4
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.7.0
fsspec==2025.5.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.2
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.15
keras==3.9.0
kiwisolver==1.4.8
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-text-splitters==0.3.8
langsmith==0.4.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mcp==1.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
namex==0.0.8
narwhals==1.31.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.1.3
openapi-pydantic==0.5.1
opt_einsum==3.4.0
optree==0.14.1
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
plotly==6.2.0
prompt_toolkit==3.0.50
propcache==0.3.2
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.1
pywin32==309
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-rst==1.3.1
rich-toolkit==0.14.8
rpds-py==0.23.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==5.0.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.4.1
stack-data==0.6.3
starlette==0.37.2
streamlit==1.43.2
sympy==1.14.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.0
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.30.1
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
xgboost==3.0.0
yarl==1.20.1
zstandard==0.23.0
kaleido
